Move `_contentToString` outside mirai-core main sourceSets, and rename it to `structureToString`
